Initial checking of objstat support code
These scripts automate running and uploading the standard objstat jobs:

objstat.sh and objstat-job.sh:
The objstat.sh script runs the 4 "standard" jobs in parallel based on
the given crawl directory (default of cwd) and number of iterations. The
jobs are described in the README at this url, which also has the latest
data:
https://drive.google.com/drive/folders/0B7VXhHzhWWb7S282VWhLVWRXbG8

UploadObjstat:
Google script and web interface to upload the zip files created by
objstat.sh and automatically create corresponding Google sheets that are
properly formatted. Looks for a google drive folder ObjStat in which to
place them. This requires a google drive account and allowing the
application to access to your google settings.